[Hosting] Register FileImageSourceService as keyed service
With keyed services introduced in .NET 8, the complexity of registering ImageSource services
can be reduced. Consequently, trimming unfriendly and AOT unfriendly code can be removed.

The idea here is to register the various ImageSource services as a keyed service to the MauiApp Service Collection
using the ImageSource type as the service key.